Remove ProviderCard
